move the public dataset v4 endpoints in own controller
